Update settings.ts
i noticed while prepping my nations (1000+ of them to be exact) that the settings page took 5+ seconds to load, freezing the entire page in the meantime. by using a documentfragment instead of repeatedly directly modifying the DOM's innerHtml, i get a small stutter when loading the settings page with 1000+ nations loaded instead of the page freezing. i also applied the same logic to the switchers display for consistency's sake.